requirements.txt
Django==1.7.1
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
ParsePy==0.3.20141108
Pillow==2.6.1
Werkzeug==0.9.6
gunicorn==19.1.1
image==1.3.4
itsdangerous==0.24
six==1.8.0
wsgiref==0.1.2